Write CIAO complient fits header for Chandra simulations - step I
CIAO expects a lot if fits header info to make the data model work, which
group columns to coordinates, includes observatory keywords etc.
This comit makes a large step in that direction, but it is not debugged, tested or
complete by any means.
In this commit I also started debuggin the asol file generation (which needs those headers).