clean up e_z <-> e_y issues in gratings
commit 3ba95601e9a17b681e9681987a05f030abdf42b3
Changed the default dispersion direction to e_y in the global coordinate system.
This required changes in the coordiante system used in the dispersion equation,
which incidentially turned a right-handed int oa left-handed coordinate system.
A lot of sign error in the direction of positive/negative orders are the result.
Those where not caught, because expected values in the tests were not updated
and tests did not include enough variability in the nuiscance direction
(rays with a commponent alnog the grooves).