Fix handling of probabilities
A bug was dicovered by comparing MARXS simulations against EXCEL spreradsheets
by Randall Smith, where MARXS would give a probability for each photon that is too high.

It turns out that some fitler elements did not multpily the input probability with
the filtercurce, but replaced it with the filter curve.
The fix that that is, in essence to replace "p = p_elem" with "p *= p_elem".

This commit also add several regession tests for filters and other optical elements to
prevent this kind of mistake in the future.