Only copy docstring if not None on original function
I run into a bug like this:

class SPOModule(RectangleAperture):
    _plot_mayavi = FlatOpticalElement._plot_mayavi

Since _plot_mayavi does not have a docstring, the metaclass tries
to copy in the docstring from FlatOpticalElement._plot_mayavi
(which is None).
However, that docstring is not writable on calss creation.
This whole problem can be circumvented if docstring only get
copied if they are not None:
In that case, this copy will never be tried, because either _plot_mayavi
already has a docstring or FlatOpticalElement._plot_mayavi does not
have one either.