Fix polarization vectors
In the old schme polarization vectors were just parallel transported when reflecting off
a surface and the pobability was adjusted according to their direction.
However, if the probability for s and p polarized photons is different (e.g. for the
Bewster angle the probability for p reflected photons must indeed be 0), then
you might end up with a polarization vector that contains both s and p polarized
components, even though the probability to reflect p photons should be 0.

In this fix, polarizaiton vectors are modified when reflected.