remove sphinx from conda_dependencies to avoid version conflict
The tests failed because of an issue explained by bsipocz on the astropy-dev
mailing list.
This should be fixed with 2.0, the workaround until then is to stick with sphinx <1.6. This limitation is taken care of in ci-helpers for docs builds. At least in theory, that somehow doesn't seem to work in your case, but that brings me to this other point that you shouldn't really list sphinx in CONDA_DEPENDENCIES as its installation is taken care of automatically for the docs build and otherwise it's unused in the tests.