Revamp display dict
Replace display dicts with a version that looks up missing items in
geometry. In this way, useful defaults can be taken fomr the geometry
(e.g. the shape) withou actually making a copy of the data (which
risks not being up to date in something is changed in the geometry).

This works for now and is a step forward, but might still evolve further
before the next release, e.g.
- use traitlets to allow default for color
- copy reference to value into display dict, taking away the need to define
  a special class.

These options need to be investigated.