Remove tests comparing the MARX C code to the Python version of HRMA
Due to Chandra-MARX/marx/#21 the C implementation and the Python
implementation of the aperture are not longer directly comparable.
For the simple version of a perfect thin lens in Python the aperture
sits right on top of the mirror plane and this the issue in
Chandra-MARX/marx/#21 is not relevant here.
As a consequence, aperture sizes and photon probabilities are no
longer directly comparable and those tests need to be removed.