Set position of hinge point where HETG is swung in and out of the beam
I have not found a true engineering drawing of this part of Chandra,
because this part is outside of the scope of the HETG. However, I did
study die HETG CSR which has a detailed 3D image of the machanism mith
axes marked, so I could deduce which of the three HETG attachment points
is closed to the Y acis. Then, I can see on the HETG drawings which
secter is closest to that. On the other hand, the radius is an educated
guess based on that 3D drawing and the known radius of the HESS.