Fix bug in SphericalDiskSource
This has nothing do do with the rest of the commit and I do not understand why the tests
did not catch this before. I suspect that an update in scipy around the time this PR
was done changed the bahavior of the function I used in the test.
Because other commits update the travis configuration to make test work again, this would
require considerable work to isolate, thus I include this fix with the current work in [#202].