Add configuration options for the precision at which color and coordinates are written to text files.
The purpose is to use rounding to reduce the file size of the XML/json/any file that represents floats as ASCII output and thus speed up loading times.

This commit only makes use of them in X3D, as the threejs backend is not that well maintained and needs to be overhauled anyway.